updating of upload/download dynamic
Several small changes made to the way in which files are uploaded and
downloaded to reduce the pressure on the server. This represents the
full project as it stands and should be runnable by all members if
nodejs and mongo are installed. NB, not branched from Sylvester's branch
directly as minor unpushed changes had to be made before-hand.